+++ packages/wiz-ui-react/src/components/base/inputs/search-input/components/types.ts
@@ -1,5 +1,10 @@
+export type Tag = {
+  label: string;
+};
+
 export type SearchInputOption = {
   label: string;
   value: number;
+  tag?: Tag;
   children: SearchInputOption[];
 };