+++ packages/wiz-ui/src/components/base/inputs/search-input/search-popup.vue
@@ -21,6 +21,7 @@
             v-for="(item, key) in option.children"
             :key="`${item.label}_${item.value}_${key}`"
           >
+            <!-- Dropdown -->
             <div
               v-if="item.children.length"
               :class="searchPopupDropdownItemStyle"
@@ -31,53 +32,135 @@
                   selectedItem.includes(item.value) &&
                     searchDropdownSelectingItemStyle,
                 ]"
+                style="
+                   {
+                    align-items: center;
+                    justify-content: space-between;
+                  }
+                "
                 @mouseover="onMouseover(item.value, option.children)"
                 @mouseout="activeItem = null"
               >
-                {{ item.label }}
-                <WizIcon
-                  size="xl2"
-                  :icon="WizIChevronRight"
-                  :color="computedIconColor(item.value)"
-                />
+                <!-- Dropdown(Tagつき) -->
+                <template v-if="item.tag">
+                  <WizHStack width="100%" justify="between" align="center">
+                    <div style="width: calc(100% - 4.5rem)">
+                      {{ option.label }}
+                    </div>
+                    <WizHStack gap="xs" width="70px">
+                      <span :style="{ lineHeight: THEME.fontSize.sm }">
+                        <WizTag
+                          :label="item.tag.label"
+                          variant="white"
+                          width="20px"
+                          font-size="xs2"
+                        />
+                      </span>
+                      <WizIcon
+                        size="xl2"
+                        :icon="WizIChevronRight"
+                        :color="computedIconColor(item.value)"
+                      />
+                    </WizHStack>
+                  </WizHStack>
+                </template>
+                <!-- Dropdown(Tagなし) -->
+                <template v-else>
+                  {{ item.label }}
+                  <WizIcon
+                    size="xl2"
+                    :icon="WizIChevronRight"
+                    :color="computedIconColor(item.value)"
+                  />
+                </template>
               </div>
             </div>
+            <!-- Checkbox -->
             <div
               v-else
               :class="searchDropdownCheckboxItemStyle"
               @mouseover="activeItem = item.value"
               @mouseout="activeItem = null"
             >
-              <input
-                v-model="checkValues"
-                :value="item.value"
-                :class="searchCheckboxInputStyle"
-                type="checkbox"
-                :id="`${item.label}_${item.value}`"
-                :name="`${item.label}_${item.value}`"
-              />
-              <label
-                :class="[
-                  searchCheckboxLabelStyle,
-                  (checkValues.includes(item.value) ||
-                    activeItem === item.value) &&
-                    searchCheckboxLabelCheckedStyle,
-                ]"
-                :for="`${item.label}_${item.value}`"
-              >
-                <WizICheck
-                  v-if="checkValues.includes(item.value)"
-                  :class="searchCheckboxIconStyle"
+              <!-- Checkbox(Tagつき) -->
+              <template v-if="item.tag">
+                <WizHStack width="100%" justify="between" align="center">
+                  <div style="width: `calc(100% - 2.5rem) `">
+                    <input
+                      v-model="checkValues"
+                      :value="item.value"
+                      :class="searchCheckboxInputStyle"
+                      type="checkbox"
+                      :id="`${item.label}_${item.value}`"
+                      :name="`${item.label}_${item.value}`"
+                    />
+                    <label
+                      :class="[
+                        searchCheckboxLabelStyle,
+                        (checkValues.includes(item.value) ||
+                          activeItem === item.value) &&
+                          searchCheckboxLabelCheckedStyle,
+                      ]"
+                      :for="`${item.label}_${item.value}`"
+                    >
+                      <WizICheck
+                        v-if="checkValues.includes(item.value)"
+                        :class="searchCheckboxIconStyle"
+                      />
+                      <span
+                        :class="[
+                          (checkValues.includes(item.value) ||
+                            activeItem === item.value) &&
+                            searchCheckboxBlockCheckedStyle,
+                        ]"
+                        >{{ item.label }}</span
+                      >
+                    </label>
+                  </div>
+                  <span :style="{ lineHeight: THEME.fontSize.sm }">
+                    <WizTag
+                      :label="item.tag.label"
+                      variant="white"
+                      width="20px"
+                      font-size="xs2"
+                    />
+                  </span>
+                </WizHStack>
+              </template>
+
+              <!-- Checkbox(Tagなし) -->
+              <template v-else>
+                <input
+                  v-model="checkValues"
+                  :value="item.value"
+                  :class="searchCheckboxInputStyle"
+                  type="checkbox"
+                  :id="`${item.label}_${item.value}`"
+                  :name="`${item.label}_${item.value}`"
                 />
-                <span
+                <label
                   :class="[
+                    searchCheckboxLabelStyle,
                     (checkValues.includes(item.value) ||
                       activeItem === item.value) &&
-                      searchCheckboxBlockCheckedStyle,
+                      searchCheckboxLabelCheckedStyle,
                   ]"
-                  >{{ item.label }}</span
+                  :for="`${item.label}_${item.value}`"
                 >
-              </label>
+                  <WizICheck
+                    v-if="checkValues.includes(item.value)"
+                    :class="searchCheckboxIconStyle"
+                  />
+                  <span
+                    :class="[
+                      (checkValues.includes(item.value) ||
+                        activeItem === item.value) &&
+                        searchCheckboxBlockCheckedStyle,
+                    ]"
+                    >{{ item.label }}</span
+                  >
+                </label>
+              </template>
             </div>
             <WizDivider
               v-if="key !== option.children.length - 1"
@@ -100,7 +183,7 @@
 </template>
 
 <script setup lang="ts">
-import { ComponentName } from "@wizleap-inc/wiz-ui-constants";
+import { ComponentName, THEME } from "@wizleap-inc/wiz-ui-constants";
 import {
   searchCheckboxBlockCheckedStyle,
   searchCheckboxIconStyle,
@@ -118,7 +201,13 @@ import {
 } from "@wizleap-inc/wiz-ui-styles/bases/search-input.css";
 import { PropType, computed, ref, watch } from "vue";
 
-import { WizDivider, WizIcon, WizSearchPopup } from "@/components";
+import {
+  WizDivider,
+  WizHStack,
+  WizIcon,
+  WizSearchPopup,
+  WizTag,
+} from "@/components";
 import { WizICheck, WizIChevronRight } from "@/components/icons";
 
 import { SearchInputOption } from "./types";