+++ packages/wiz-ui-next/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @wizleap-inc/wiz-ui-next
 
+## 2.11.0
+
+### Minor Changes
+
+- [#1036](https://github.com/Wizleap-Inc/wiz-ui/pull/1036) [`c15617e2`](https://github.com/Wizleap-Inc/wiz-ui/commit/c15617e2b10300f6fe79293580edff6d13684a03) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Feat(WizTag): Variant を追加
+  Feat(WizSearchInput): tag 機能を追加
+  Feat(WizStack, WizHStack, WizVStack): className, style を公開
+  Feat(WizCheckBox): className, style を公開
+
+### Patch Changes
+
+- Updated dependencies [[`c15617e2`](https://github.com/Wizleap-Inc/wiz-ui/commit/c15617e2b10300f6fe79293580edff6d13684a03)]:
+  - @wizleap-inc/wiz-ui-styles@0.21.0
+
+## 2.10.0
+
+### Minor Changes
+
+- [#1068](https://github.com/Wizleap-Inc/wiz-ui/pull/1068) [`4ec8a235`](https://github.com/Wizleap-Inc/wiz-ui/commit/4ec8a235a19eabd3859f0a0d9b982b557acc4060) Thanks [@wmoai](https://github.com/wmoai)! - WizRadioNew を追加
+
+- [#1067](https://github.com/Wizleap-Inc/wiz-ui/pull/1067) [`888eb872`](https://github.com/Wizleap-Inc/wiz-ui/commit/888eb872237b9f4ce66b6e333eac43de47c886a3) Thanks [@wmoai](https://github.com/wmoai)! - WizCheckBoxNew を追加
+
+- [#1051](https://github.com/Wizleap-Inc/wiz-ui/pull/1051) [`d74864b4`](https://github.com/Wizleap-Inc/wiz-ui/commit/d74864b4f97092693d343dba4a991cbf40b22be6) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Fix(button): button 要素に type を付与
+
+### Patch Changes
+
+- Updated dependencies [[`4ec8a235`](https://github.com/Wizleap-Inc/wiz-ui/commit/4ec8a235a19eabd3859f0a0d9b982b557acc4060), [`888eb872`](https://github.com/Wizleap-Inc/wiz-ui/commit/888eb872237b9f4ce66b6e333eac43de47c886a3)]:
+  - @wizleap-inc/wiz-ui-constants@0.14.0
+  - @wizleap-inc/wiz-ui-styles@0.20.0
+
+## 2.9.0
+
+### Minor Changes
+
+- [#1049](https://github.com/Wizleap-Inc/wiz-ui/pull/1049) [`dc602da7`](https://github.com/Wizleap-Inc/wiz-ui/commit/dc602da7d1ccdb55972f30ef06520f0b9870e730) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Feat(Icon): 新規アイコン追加
+  Chore(IconGen): Vue2 サポートを停止
+
+- [#1018](https://github.com/Wizleap-Inc/wiz-ui/pull/1018) [`795783e4`](https://github.com/Wizleap-Inc/wiz-ui/commit/795783e4a52675d6830c0a86de602feca7fddee2) Thanks [@okakyo](https://github.com/okakyo)! - [#1017] Snackbar に expand property を追加
+
+### Patch Changes
+
+- Updated dependencies [[`dc602da7`](https://github.com/Wizleap-Inc/wiz-ui/commit/dc602da7d1ccdb55972f30ef06520f0b9870e730), [`795783e4`](https://github.com/Wizleap-Inc/wiz-ui/commit/795783e4a52675d6830c0a86de602feca7fddee2)]:
+  - @wizleap-inc/wiz-ui-constants@0.13.0
+  - @wizleap-inc/wiz-ui-styles@0.19.0
+
 ## 2.8.0
 
 ### Minor Changes